# it-helpdesk-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IT Helpdesk API",
    "description": "사내 IT 헬프데스크 시스템 REST API - 티켓 생성, 조회, 업데이트 기능 제공",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "<HelpdeskUrl>",
      "description": "Production (Route53 → ALB → ECS)"
    }
  ],
  "paths": {
    "/tickets": {
      "get": {
        "summary": "티켓 목록 조회",
        "description": "모든 티켓을 조회하거나 필터링하여 조회합니다",
        "operationId": "getTickets",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "상태로 필터링 (open, in_progress, resolved, closed)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["open", "in_progress", "resolved", "closed"]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "우선순위로 필터링 (low, medium, high, urgent)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "urgent"]
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "카테고리로 필터링",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["hardware", "software", "network", "account", "other"]
            }
          },
          {
            "name": "requester",
            "in": "query",
            "description": "요청자 이름으로 검색",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "티켓 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "새 티켓 생성",
        "description": "새로운 IT 헬프데스크 티켓을 생성합니다",
        "operationId": "createTicket",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "티켓 생성 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "summary": "티켓 상세 조회",
        "description": "특정 티켓의 상세 정보를 조회합니다",
        "operationId": "getTicket",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "티켓 ID (예: TICKET-0001)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "티켓 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "404": {
            "description": "티켓을 찾을 수 없음"
          }
        }
      },
      "patch": {
        "summary": "티켓 업데이트",
        "description": "티켓의 상태, 담당자, 우선순위를 업데이트합니다",
        "operationId": "updateTicket",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "티켓 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "티켓 업데이트 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "404": {
            "description": "티켓을 찾을 수 없음"
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "티켓 통계 조회",
        "description": "전체 티켓 통계를 조회합니다 (상태별, 우선순위별, 카테고리별)",
        "operationId": "getStatistics",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "티켓 통계",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "전체 티켓 수"
                    },
                    "by_status": {
                      "type": "object",
                      "description": "상태별 티켓 수"
                    },
                    "by_priority": {
                      "type": "object",
                      "description": "우선순위별 티켓 수"
                    },
                    "by_category": {
                      "type": "object",
                      "description": "카테고리별 티켓 수"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Helpdesk REST API 호출 시 필요한 API 키"
      }
    },
    "schemas": {
      "TicketCreate": {
        "type": "object",
        "required": ["title", "description", "category", "requester"],
        "properties": {
          "title": {
            "type": "string",
            "description": "티켓 제목",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "상세 설명",
            "minLength": 1
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "urgent"],
            "default": "medium",
            "description": "우선순위"
          },
          "category": {
            "type": "string",
            "enum": ["hardware", "software", "network", "account", "other"],
            "description": "카테고리"
          },
          "requester": {
            "type": "string",
            "description": "요청자 이름",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
      "TicketUpdate": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["open", "in_progress", "resolved", "closed"],
            "description": "티켓 상태"
          },
          "assigned_to": {
            "type": "string",
            "description": "담당자"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "urgent"],
            "description": "우선순위"
          }
        }
      },
      "Ticket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "티켓 ID"
          },
          "title": {
            "type": "string",
            "description": "티켓 제목"
          },
          "description": {
            "type": "string",
            "description": "상세 설명"
          },
          "status": {
            "type": "string",
            "enum": ["open", "in_progress", "resolved", "closed"],
            "description": "티켓 상태"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "urgent"],
            "description": "우선순위"
          },
          "category": {
            "type": "string",
            "enum": ["hardware", "software", "network", "account", "other"],
            "description": "카테고리"
          },
          "requester": {
            "type": "string",
            "description": "요청자 이름"
          },
          "assigned_to": {
            "type": "string",
            "nullable": true,
            "description": "담당자"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "생성일시"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "수정일시"
          }
        }
      }
    }
  }
}